Day5.EDSL

Documentation

data Crate

A crate with a character in it, or an empty crate (only used for parsing)

Constructors

Crate Char

EmptyCrate

data Move

A move with the amount of boxes, a location from where to pick the crates and where to move them to.

Constructors

Move Int Int Int

newtype Stack

Constructors

Stack [Crate]