Day7.Evaluator

Documentation

evalTree :: Tree -> Tree

Evaluate the size of all folders and add to each Node its size

calculateSum :: Tree -> Int

Traverse the tree and add all folders where the size is at most 100000 Ignore single files

findMin :: Tree -> Int

Find the minimal size of a folder to delete from a given tree Assume the root node of the tree to be the entire filesystem