Day7.TreeBuilder

Documentation

buildTree :: [Command] -> Tree

Build a tree from a list of commands Utilizes the list of commands to traverse and fill the tree

data Tree

Data structure storing the folders as nodes and files as trees

Constructors

Node String Int [Tree]

Leaf String Int