Day8.Day8

Documentation

getDaySolutions :: (String -> String, String -> String)

scenicScore :: [Int] -> Int

Calculate the scenic score in one direction for a given list of numbers The scenic score of the first element x in the list is all values less than x, up to and including the first element larger or equal to x The scenic score of a value with no other elements next to it is 0

scenicScores :: [Int] -> [Int]

Calculate the scenic scores for all elements in a given list